fix parsing defcomputedcolumn
This puts through some minor fixes for the parser in relation to
defcomputedcolumn.  It also enables the ComputedRegister assignment to
be encoded using the gob format (i.e. so it can be included in the bin
file).  Finally, I removed this spurious debug binary --- not sure what
that was for?