feat: improve formatted lisp output (#1065)
* improved reported for power-of-2 constants

This puts in place improved lispification for constants which are
powers-of-2.  This just improves readability in certain circumstances,
especially when we have lots of register splitting.

* use "if" in place of "ite" for lisp output

This favours using "if" rather than the (confusing) "ite" for lisp
output.

* simplify constraints generated from assembly

Constraints were not being simplified when being generated from assembly
files, and this resulted in output which was less than ideal.

* formatting is improved, but not perfect

There are still lots of problems.

* improve formatting

This puts through a number of improvements to formatting, such as only
showing fully qualified names in global scopes.  Also, improvements to
the formatting algorithm (though more work is needed here for sure).