working on framing for multiline functions
There are several awkward issues to resolve.  Firstly, we have to
introduce the return line somehow; secondly, it makes sense to have the
return line and pc as part of the function registers, but only when we
have multi-line functions (compared with the moment, where the PC is
always included).  To that end, having the PC at the end of the register
file might also make more sense.