feat: fix control flow book keeping (#1080)
* initial draft of Framining concept

The concept of a framing is that it abstracts the various bookkeeping
needed for representing a function in a trace.  For example, control
lines needed to identify a terminal state in a multi-line function.

* working on framing for multiline functions

There are several awkward issues to resolve.  Firstly, we have to
introduce the return line somehow; secondly, it makes sense to have the
return line and pc as part of the function registers, but only when we
have multi-line functions (compared with the moment, where the PC is
always included).  To that end, having the PC at the end of the register
file might also make more sense.

* remove pc from function registers

Previously a special (and essentially unused) register was added at
index 0 of a function's registers.  This doesn't really make sense to my
mind.

* multi-line functions now have control lines

The control lines are being computed as part of trace expansion for
multi-line functions.  At this stage, however, the proper set of
constraints are not.  Likewise, padding is not being set properly either
way for witnesses.

* add constraints for multi-line functions

What remains is to handle padding for single-line functions.

* fix: encoding / decoding for tables

* minor bug fixes related to PC register

There were some minor bug fixes which carried over from removing the PC
register at the UASM level.

* fixes for padding

The key challenge is that lookups into a ZK function automatically have
the initial padding row inserted.