feat: enable recursive definition for a defcomputedcolumn (#1088)
* add tests / updated parser

This adds some tests for recursive computed columns, and updates the
parser to support the annotations ":fwd" and ":bwd".

* Recursive computation now operational

* detect recursive definitions

This adds support for detecting and report errors for recursive
definitions.

* mark tests as ignored (for now)

These tests cannot be resolved without further work, and have been
marked as ignored (alongw with issue number).

* fix ModuleScope.IsVisible()

This was not sufficiently flexible to account for all lookup scenarios.
Instead, have copied the essential structure of Bind() to get this to
work correctly.