outstanding issue with padding
A key issue is that the padding computation in ComputedRegister was
removed because we no longer return a tr.ArrayColumn, but instead just
return an array.MutArray.  To work around this, we need to figure out
appropriate padding values.

This puts in place preliminary support for padding declarations at the
source level, and within the schema.

This allows padding values to be correctly determined for normalisation
and expansion gadgets.