feat: transition away from field array implementations (#1097)
* embed word pool within lt.TraceFile

This stores the word pool within the TraceFile itself.  This should
(hopefully) enable reduce times for splitting.  It is also a precursor
to removing the duplicate code between the field and array packages.

* refactor many things to be field generic

This puts through a large number of changes which makes many things
field generic.  However, it stops short of making everything field
generic purely to keep the size of this change somewhat in hand.  The
expectation is that there will be further steps required to fully
realise generic fields.

There are some subtle aspects around the issues of representing bytes in
an encoded form for field elements.

* static padding

A key issue is that the padding computation in ComputedRegister was
removed because we no longer return a tr.ArrayColumn, but instead just
return an array.MutArray.  To work around this, we need to figure out
appropriate padding values.

This puts in place preliminary support for padding declarations at the
source level, and within the schema.  This allows padding values to be 
correctly determined for normalisation and expansion gadgets.

* traces now hold words, not field elements

This puts through a fairly significant refactoring of the previous
approach whereby traces only hold words rather than field elements.
Instead, field elements are instantiated on demand as required.  This is
quite a nice solution since most of the work involved in trace expansion
is really about manipulating words.  The trace wrapper is responsible for 
making a trace of words look like a trace of field elements.  This also adds
 a field array which wraps an existing  word array.  This is only useful in a 
few situations where we want to see the  world in terms of fields.

* update tests

This updates the memory benchmark tests by significantly expanding the
set of accepts and reejects tests based on randomly generated memory
read/write sequences.

This adds specific tests for the memory16u8 example to ensure that it
tests different bitwidth columns which, in turn, ensure different column
implementations are tested.