fix: representation of column data (#1101)
* represent trace as columns of field elements

This restores the original strategy of storing the trace as a collection
of concrete field elements, rather than as a collection of dynamic words
which are then converted into field elements on demand.  This should
improve performance.

* refactoring to reduce direct use of word.Word

Instead, field.Element[F] is preferred or word.BigEndian.

* add StaticPool and DynamicWord

The idea is that a DynamicWord corresponds with word.BigEndian, in that
it contains a variable number of bytes.  Methods specifically needed for
that are then included in this interface only.

StaticPool differs from HeapPool in that it stores whole words, rather
than compressing them into short byte sequences.

* bugfix for legacy byte decomposition

* fix computation bug