fix: improve memory and CPU performance (#1108)
* refactor word array package

This puts through quite a few refactorings for the word array package,
specifically in an attempt to reduce memory footprints.

* fix index pools

This puts through a simple fix for index pools which ensures they always
have at least one element for the zero index.

* various attempts to improve things

Performance seems to be good for checking individual traces, but not
good for running tests.

* ensure small pools are really static

The small were causing significant overhead for the tests because they
were begin freshly allocated for each test.