feat: finalise generic fields (#1118)
This updates components of the system to support some target field F.
For example, the top-levelfunction mir.Concretize() is responsible for
converting a schema in the Mid-level Intermediate Representation (MIR)
from one field (F1) into another field (F2).  This is a necessary step
towards supporting small fields.  The intuition is that the schema
starts in some neutral word format (e.g.  uint256) and is then lowered
into a specific field (e.g. KoalaBear).

The concretization process also mostly replaces the existing
FieldAgnostic interface, which was limited due to the need for source
and target fields.  Instead, we concretize at the assembly level as
this level is all knowing.

The various CLI commands, and the testing framework, are also updated
to support a fixed number of fields.  Currently, we support GF_251,
GF_8209, KoalaBear and BLS12_377.  More fields can be added easily
enough, but it requires updating the instance lists for each command.
The target field is determined by the '--field' CLI argument.

NOTES: Having updated golangci-lint version, there were a number of
outstanding lint errors to fix.