add erroneous tests
This adds a number of tests which should not fail, but currently are.
The reason is related to the occurence of logical false within a
constraint.  Currently, this is not being handled properly.

The observations for these tests arose because the fail instruction
results in logical false being inserted into the constraints.