feat: support fail instruction with padding inference (#1129)
* support fail instruction with padding inference

This puts in place support for the fail instruction, along with a
straightforward notion of padding inference.  However, padding inference
is only guaranteed to succeed when the "null input" for a function is a
valid instance.

* add erroneous tests

This adds a number of tests which should not fail, but currently are.
The reason is related to the occurence of logical false within a
constraint.  Currently, this is not being handled properly.

The observations for these tests arose because the fail instruction
results in logical false being inserted into the constraints.

* fix for logical false

There were some issues related to logical false being translated
incorrectly.  Should be fixed now.

* support optional padding specifiers

This adds support for (optional) padding specifiers on the parameters of
assembly functions.  This allows one to effectively provide a "witness"
which is always used within the padding region.

* fix trace expansion in presence of failure

This tweaks the way that trace expansion reacts when execution
encounters a fail instruction.  Previously, the code simply raised a
panic which meant that reject tests were not executing correctly.
Instead, the erroneous traces are simply returned as is on the
understanding they should fail any generated constraints.