feat: implement ltv2 file format (#1137)
* basic encoding / decoding operational

This puts in place a straightforward implementation of the ltv2 file format, which allows a greater degree of compression than the original format.

* support small arrays

This puts in place support for small arrays, rather than pool arrays as
were previously used (which was not efficient).

* Optimise column widths

In some cases, we can optimise column widths and this can make a
reasonable difference.  However, there is still more which can be done
here for sure.