fix: lowering macrocall (#1140)
* fix macro.Call.Lower

A simple bug in this method.

* refactor set.Comparable

This refactors the set.Comparable interface so that it doesn't require
the underlying types to be comparable.

* implement Executor for InferPadding

The executor enables efficient execution of a program by caching
previously seen function inputs.  This is needed to correct determine
padding values in the presence of function calls.  However, it remains
to use the executor for trace expansion.

* minor fixes