support function calls (without coordination)
This reinstates support for filling traces involving function calls.
This uses a common executor to allow assignments to determine expected
return values for function calls.  However, it falls short of
implementing a full coordination phase, meaning that traces must start
with all required function instances.

An essential part of the mechanism allowing this to work is an API which
sits inbetween asm programs and MIR schemas.  This is a wrapper layer
whose top-level construct are the MixedProgram[T] instances.  These
replace the previous MixedSchema[T] instances, etc.