refactor lt.TraceFile layout
This refactors the lt.TraceFile layout to be organised by module, rather
than the previous flat structure.  Thus, each module contains zero or
more columns, etc.  This will work well with the new ltv2 format, but
was a fairly significant change in the code base.

The primary reason for making this change at this time is to facilitate
"trace alignment" as a separate task which occurs before trace
propagation and trace building.