implement trace alignment
Trace alignment is the process of ensuring the layout of modules in the
schema and trace match as expected.  The process fails, for example, if
the trace is missing a module or a column, etc.  This is, in part, a
safety check to ensure the trace is well-formed.  However, it is also
necessary for correct operation of the system.