implement trace propagation
trace propagation is the process of propagating secondary function
instances throughout the call graph, based on those primary instances
provide in the original trace.  Specifically, suppose a function "f(x)"
calls a function "g(y)".  Suppose the trace contains one instance, say
"f(1)=3", but has nothing for "g(y)".  Then, the purpose of trace
propagation is to determine what instance is needed for "g(y)" based on
the execution of "f(1)", and add it to the trace.