fix: reinstate recursive assembly tests (#1141)
* support function calls (without coordination)

This reinstates support for filling traces involving function calls.
This uses a common executor to allow assignments to determine expected
return values for function calls.  However, it falls short of
implementing a full coordination phase, meaning that traces must start
with all required function instances.

An essential part of the mechanism allowing this to work is an API which
sits inbetween asm programs and MIR schemas.  This is a wrapper layer
whose top-level construct are the MixedProgram[T] instances.  These
replace the previous MixedSchema[T] instances, etc.

* sketch trace propagation outline

Trace propagation is the process of propagating secondary function
instances throughout a trace.  This is one in such a way as to ensure no
two identical instances are generated.  Furthermore, this is done prior
to trace splitting since it is more efficient at this point.

* refactor lt.TraceFile layout

This refactors the lt.TraceFile layout to be organised by module, rather
than the previous flat structure.  Thus, each module contains zero or
more columns, etc.  This will work well with the new ltv2 format, but
was a fairly significant change in the code base.

The primary reason for making this change at this time is to facilitate
"trace alignment" as a separate task which occurs before trace
propagation and trace building.

* implement trace alignment

Trace alignment is the process of ensuring the layout of modules in the
schema and trace match as expected.  The process fails, for example, if
the trace is missing a module or a column, etc.  This is, in part, a
safety check to ensure the trace is well-formed.  However, it is also
necessary for correct operation of the system.

* implement trace propagation

trace propagation is the process of propagating secondary function
instances throughout the call graph, based on those primary instances
provide in the original trace.  Specifically, suppose a function "f(x)"
calls a function "g(y)".  Suppose the trace contains one instance, say
"f(1)=3", but has nothing for "g(y)".  Then, the purpose of trace
propagation is to determine what instance is needed for "g(y)" based on
the execution of "f(1)", and add it to the trace.

* reinstate recursive assembly tests

This puts back in place the three recursive assembly tests that were
previously failing due to a lack of trace propagation.  These tests
require trace propagation to complete their traces by filling out all
secondary function instances.

* postpone Shift and RecPow tests

These recursive tests fail for some reason related to padding.  This may
be connected with the (lack of) conditional lookups.  Since it is
separate from the main goal of this feature, it is deferred.