fix: legacy reader and writer
This puts through two separate fixes for the legacy reader and the
legacy writer.  The latter was a simple issue whereby an incorrect
number of bytes were being written for variable sized word.BigEndian
values.  The legacy reader issue was, unfortunatly, more fundamental and
related to the change in representation of the lt.TraceFile.
Specifically, how it is now organised by module.  The implementation
attempted to read and organise at the same time, but was completely
broken.  Now, the original code is retained for reading with
organisation occuring after the fact.