require assignments to support substitution
Whilst Constraint instances were already required to support
substitution, assignments were not.  This perhaps made sense previously
when there was no way for a source-level expression to end up in an
assignment before substitution was applied on it.  However, assignments
can now be written directly within the source, and this opens up the
possibility that an expresion in an assignment was not substituted when
it should have been.