fix: native interleaving bitwidth calculation (#1157)
The calculation of the target bitwidth for the native interleaving
comptuation incorrectly assumed that bitwidths must match.  In fact,
bitwidths in the dynamic arrays do not have to match that of their
corresponding register.  For example, if a column contained only 0 or 1
we could implement it using a bit array, regardless of whether its
register was a u8 or a u16, etc.

This fix simply updates the calculation to take the maximum bitwidth of
any source column to use for the target column.  This is safe, and
optimal.