add enable line to io.Bus
This adds a notion of an enable line to an io.Bus.  This allows the bus
to be enabled or disabled, as required.  However, at this stage, the
enable line is never filled during trace expansion with anything other
than 0.  Furthermore, no constraints are added to enforce its one at
certain positions.