enforce valid trace instance for all functions
This puts in place a change of approach, whereby the trace is now
expected to have at least one valid instance of every function.  This,
in turn, means we can use unconditional lookups (in the source position)
for implementing function calls.  This is because, when the bus is not
in use, it can always be filled with the given valid instance.