fix: failing lookup for inactive function call (alternate solution) (#1161)
* prevent constancy constraints for I/O lines

* enforce valid trace instance for all functions

This puts in place a change of approach, whereby the trace is now
expected to have at least one valid instance of every function.  This,
in turn, means we can use unconditional lookups (in the source position)
for implementing function calls.  This is because, when the bus is not
in use, it can always be filled with the given valid instance.

* employ conditional lookups to check return lines

This employs a conditional lookup for multi-line functions, where by the
target selector is set as the return line.  Thus, only when this line is
1 will the target row be used within the lookup.  This means we cannot
lookup part way into a function instance.