initial implementation of exp test
This puts in place a partial implementation of the exp test which passes
a small number of tests.  There are a few things about this
implementation.  Firstly, it uses an inefficient log256 implementation
(i.e. non-recursive).  Secondly, it only provides an implementation for
the EXPLOG case, but not yet the MODEXPLOG case.