feat: support checked lookups (#1165)
* support always-on lookup checking

This puts in place the basic mechanism for checking source/target types
in lookups.  At this stage, the mechanism is always on.  The next step
is to add an attribute which can be used to disable checking for a
lookup.

* support :unchecked lookup attribute

This adds support for the ":unchecked" lookup attribute across all three
forms of lookup, including deflookup and defclookup.

* fix existing tests

A small number of existing tests fail with the new, stricter, checking
of lookups.  This simply fixes those tests so they now pass.