feat: support include directive (#1168)
* support parsing include declarations

This adds support for parsing and accumulating include declarations.
However, at this stage, the list of includes is not processed in any
way.

* support reading included files

This adds simplistic support for reading included files.  This attempts
to prohibit duplicates from being read by exploiting go's filepath
package.  However, there are still cases where this can be fooled.  For
now, it is powerful enough.