remove shared io.Executor
This removes the io.Executor shared amongst assignments, since was not
thread-safe yet trace expansion can exploit parallelism.  Specifically,
this was leading to race conditions where the shared executor was being
read / written concurrently.

The proposed solution is to reuse the existing trace for assignment
computation.  Specifically, the trace contains all of the valid
instances which were generated during trace propagation.  As such, it
can (in principle) answer necessary queries regarding I/O instances.

There are challenges with this route though.  In particular, the modules
of a trace are being expanded concurrently, meaning trace data is not
immutable.  One option would be to make copies of the initial
(unexpanded) data.  The other solution would be to synchronise the trace
when the trace for a given module is expanded.