implement fill_bytes_from
This provides an implementation which fills all the bytes in a given
u256 word with a given value, starting from a given offset.  This is
used in the exp module to implement the "trim" functionality.

This now includes all test inputs for the exp module found in the
unit & replay tests.