perf: Integrate Batch Inversion in the PseudoInverse gadget (#1149)
* refactor: PseudoInverse class in assignment package

* feat: PeudoInverse as assignment

* refactor: clean up unnecessary functions

* fix: "inverse" -> "inv" in Lisp

* feat: use batch inverse

* refactor: rename to pseudo_inverse.go

* remove recursiveModule wrapper from PseudoInverse

This is not needed for the PseudoInverse computation, so removed for
simplicity.

* fix padding inference for pseudo-inverse columns

The padding computation for pseudo-inverse columns was not being done
correctly, as it required the old pseudoInverse expression.  Have put
that back specifically just for the padding computation (which is
perhaps frusrating, but not easy to do otherwise).

* refactor: mark EvalAt as unreachable

* revert: EvalAt is reachable

---------

Co-authored-by: DavePearce <dave01001110@gmail.com>