mostly complete shf module implementation
This mostly completes the shf module implementation by using a somewhat
ugly workaround for the register splitting issue.  At this stage, there
is one outstanding bug which is highlighted by only a single test.
Specifically, when performing sar(w,n) where n >= 256 and w is signed.
In this case, the result should be -1, not 0.