feat: add shf example (#1178)
* add implementations of shl, shr and sar.

These functions provide byte-level shifting functionality, including a
signed shift right.  Specifically, they allow a given u256 to be shifted
a certain number of bytes either to the left or to the right.

* support bit_shift example and tests

* add shf implementation

This adds a (partially working) implementation of the shf module which
handles bitwised shifting for 256bit (signed and unsigned) words.  The
implementation uses a coarse byte-level shift first, followed by a fine
bit-level shift second.

* mostly complete shf module implementation

This mostly completes the shf module implementation by using a somewhat
ugly workaround for the register splitting issue.  At this stage, there
is one outstanding bug which is highlighted by only a single test.
Specifically, when performing sar(w,n) where n >= 256 and w is signed.
In this case, the result should be -1, not 0.

* complete shf implementation

This adds support for SAR for shifts >= 256 bits by correctly accounting
for the sign bit of the argument.

* disable shf test

This test is disabled (for now) since it exposes the bug in #1172.
However, all the tests pass when run sequentially.  When #1172 is
resolved, this test will be reenabled.