finalise exp implementation
In the end, this uses shr256 rather than fill_bytes to managing the
trimming and shifting in one go.  Furthermore, there is room to improve
the assembly itself, since the code is clunky in places due to
current limitations of the ASM language (which should be lifted).
Nevertheless, the implementation works.

This also further splits out the various assembly tests into "unit",
"util" and "bench" groupings.  The latter includes all current module
implementations.