feat: add exp example (#1180)
* initial implementation of exp test

This puts in place a partial implementation of the exp test which passes
a small number of tests.  There are a few things about this
implementation.  Firstly, it uses an inefficient log256 implementation
(i.e. non-recursive).  Secondly, it only provides an implementation for
the EXPLOG case, but not yet the MODEXPLOG case.

* add log2 function

This adds a recursive implementation of the log2 function which uses a
subdivision algorithm rather than a loop.  This includes a large number
of tests generated automatically from the log256 tests.

* add set_byte function

Adds an implementation of set_byte which is test with a large amount of
test data.  This should be a workhorse going forward, as its utility
extends far beyond the exp module.  This method sets the ith byte in a
given 256bit word.

* implement fill_bytes_from

This provides an implementation which fills all the bytes in a given
u256 word with a given value, starting from a given offset.  This is
used in the exp module to implement the "trim" functionality.

This now includes all test inputs for the exp module found in the
unit & replay tests.

* finalise exp implementation

In the end, this uses shr256 rather than fill_bytes to managing the
trimming and shifting in one go.  Furthermore, there is room to improve
the assembly itself, since the code is clunky in places due to
current limitations of the ASM language (which should be lifted).
Nevertheless, the implementation works.

This also further splits out the various assembly tests into "unit",
"util" and "bench" groupings.  The latter includes all current module
implementations.