fix: race condition during propagation expansion (#1181)
* minor tweaks

* Split out FunctionTrace from Executor

This splits out a separate object for capturing a function's trace
within the executor.  The benefit of this is that we can now lock data
within an individual trace.

* implement thread-safe FunctionTrace

This puts in place a thread-safe implementation of the executor's
FunctionTrace object.  This allows locking on a function-by-function
basis (to reduce contention) and also uses a read-write lock (since we
have a cache).