fix: update exp benchmark (#1184)
* support CDS cutoff in exp module

This adds support for the case when CDS < EBS in the exp module.  This
case requires bytes to be filled with 0 between the CDS and EBS markers.

* update test traces for exp module

This significantly updates the set of traces for the exp module, based
on the full set of tests available.

* update log256 function to use subdivision method

The log256 function was previously using a loop, which was inefficient
and rather awkward.  This arose because it was implemented before trace
propagation was supported.  However, since trace propagation is now
functioning well, we can use the more efficient subdivision method.

* sample down shf.accepts tests

This samples down the shf.accepts tests which were taking far too long
to run.

* sample down tests for fill_bytes and set_byte

There were also a very large number of tests here.

* reduce max padding for assembly tests

This puts through a fix which allows less padding to be tested for
assembly tests.  This is sensible, since assembly programs only ever use
unit shifts and, hence, testing large padding values is pointless.