feat: GF251 example (#1119)
 This provides an implementation of additions and multiplications on the field GF251 for u8 elements. U8 have been chosen to be able to test all cases, even if only sampled in the final unit tests for performance reasons. 

* addition

The function sums two u8s and returns the result modulo 251. It uses register splitting and carries to wrap around 251. 

* multiplication

This function multiplies two u8s and returns the result modulo 251. For the multiplication function, we have implemented the Montgomery algorithm using R=256, co-prime with 251, and a reduce function. This algorithm shifts the modulo to the R chosen, here 256, where arithmetic can be achieved by register splitting and bit shifting. The reduce function then projects the result back to the 251 modulo.