fix: race condition in `pkg/test/util/framework.go` (#1197)
* use race detector for CI tests

This updates the CI tests to use the race detector by default, meaning
that race condition will prevent merging.

* lower point of schema cloning in testing framework

This lowers the point where the schema is cloned within the testing
framework.  The purpose of this is to ensure every trace has a unique
schema, thereby avoiding sharing of the `io.Executor` across traces.

* separtate SchemaStacker from a SchemaStack

This allows the configuration for a SchemaStack to behave in a
functional manner.  Thus, a SchemaStacker is simply a configurable
builder for a schema stack.

* separate out specific racer tests

Since using the race detector on all tests is quite expensive in terms
of time, this splits out some specific tests using the race detector.

* Fix #1198

This fixes a race condition in io.Executor where the read lock was
released slightly too early.  This additionally tweaks the set of tests
being run to improve CI performance.