feat: improve trace command (#1200)
* log time for trace propagation

This helps to make the time visible so as to consider whether it should
be further optimised.

* feat: support sorting trace columns

This puts in place support for sorting trace columns, and also improving
the performance of module summarisation.

* parallelise trace reconstruction

This is where a large amount of time is being spent within the trace
command.  Ideally, simply avoiding this step would be useful.  For now,
the aim is to make it faster.

* sequentialise trace reconstruction