fix: ordering of lhs in call (#1204)
* fix: lhs ordering for call instructions

This fixes the ordering of the left-hand side for call instructions.
Previously, it had to be written in the reverse order, which was
extremely counter-intuitive!

* update stp to fix call ordering

Previously, calls in the STP benchmark were written with their lhs in
the reverse order to work around the bug which this issue fixes.
Therefore, we must write them in the expected way!