feat: support assembly constants (#1209)
* support parsing constants

This adds support to the parser for constant declarations, but these are
not yet operational within the linker.

* support constants in assignments

This adds support for constants in assignments only, since these are
currently the only instruction forms which support general expressions.

* support labelled constants in ifgoto instructions

This adds support for labelled constants for the RHS of an ifgoto
instruction.  At this stage, we cannot do more as generalising an ifgoto
to accept an arbitrary expression is challenging with respect to
register splitting.  Therefore, what we have here is something of a
temporary measure.