feat: update `shf` benchmark (#1211)
* add instance counting to the executor

This adds a simple mechanism for counting the total number of instances
held within the executor.  The purpose of this is primarily for
debugging / profiling.

* replace byte/bit_shl/shr/sar

This replaces byte_shr, byte_shl, byte_sar with improved implementations
of bit_shr, bit_shl and bit_sar.  The improved implementations use a
decomposition approach to dramatically reduce the line count.  For an
arbitrary shift over 256bits, we now require at most 8 lines whereas
before it was atmost 32lines.