fix: translation of signed polynomials (#1214)
This fixes the translation of signed polynomials in two ways.  Firstly,
when the right-hand side of an assignment consists of only a single
register, this register can never be considered a sign bit.  Secondly,
the determination of whether or not a polynomial can be signed (i.e.
evaluated to a negative value) is shifted to the existing code (which is
more robust) contained within the agnostic package.