fix scan error reporting
Previously, there was a problem whereby the errors for an incorrect scan
query were being reporting in the wrong place.  Specifically, at a
position shifted consistenty from where it should have been.  The reason
for this is that the prompt had changed (to indicate about the row
variable $), but the offset was fixed.