update inspector default IR
This updates the default IR level of the inspector from MIR to AIR.
This is to ensure consistency with the default level for the check
command.  Specifically, a workflow where the check command produces an
error (at the default AIR level) was confusing when viewed in the
inspector (at the default MIR level).