feat: asm tests for invalid syntax (#1217)
* rework invalid testing mechanism

This completely reworks the mechanism for checking correct errors are
reported for invalid source files.

* support assembly invalid tests

This puts in place basic infrastructure for ensuring syntax errors are
reported appropriately.

* add a selection of invalid tests

This adds a relatively modest selection of invalid tests, with the
intention to add more incrementally over time.  The tests immediately
started showing up issues, so they are clearly overdue.