fix: offset error in inspector scan (#1218)
* fix scan error reporting

Previously, there was a problem whereby the errors for an incorrect scan
query were being reporting in the wrong place.  Specifically, at a
position shifted consistenty from where it should have been.  The reason
for this is that the prompt had changed (to indicate about the row
variable $), but the offset was fixed.

* update inspector default IR

This updates the default IR level of the inspector from MIR to AIR.
This is to ensure consistency with the default level for the check
command.  Specifically, a workflow where the check command produces an
error (at the default AIR level) was confusing when viewed in the
inspector (at the default MIR level).