feat: optimise conditional control flow (#1223)
* working on branch translator

The current plan is to traverse consecutive skips building up a kind of
"super skip" which can then be optimised in various ways.

* implement branch table

This puts in place what should hopefully be an optimal implementation of
branch table. This makes use of path conditions (called branches)
represented in Conjunctive Normal Form.  Thus, it allows fairly powerful
reasoning about just branches (in principle).

* complete branch table translation

This finalises the branch table translation.  At this stage, there
remain improvements which could arise in the representation of micro
instructions.  Specifically, when duplicate instruction sequences arise,
these should be collapsed.  Also, there are cases where the branch table
translation does not apply enough simplifications to produce an optimal
solution.

* improve worklist algorithm

This improves the worklist algorithm to exploit known facts of the
micro.Instruction internal representation.  Specifically, that skips
almost move to the right.

* fix for unreachable branches

* tweak memory performance

This applies some simple tweaks arising from memory profiling.

* disable use of Koalabear in places

The use of Koalabear in a few places appears to cause a pathalogical
state explosion with the new branch table algorithm.  This has something
to do with the introduction of many (presumably unoptimised)
conjunctions.

* tweak notion of contradiction

* ignore koalabear on problematic tests

Use of the Koalabear field on certain tests appears to be causing an
exponential blow up, and will need to be fixed.  For now, am attempting
to merge the feature without addressing this issue (i.e. because
Koalabear is not actually used in anger yet).

* prevent if-else for branch already considered

* seperate Proposition from BranchTable

This separates the logical proposition and its implementation from the
BranchTable implementation.  Thus, the former can be tested in isolation
and reused in other parts of the compiler as/when needed.

* support various logical simplifications

This increases the range of logical simplifications supported for branch
conditions.  However, there remains a way to go in terms of reaching an
ideal set.  Furthermore, this adds a number of logical equivalence
tests.