fix: bug in polynomial subtraction (#1234)
* fix bug in polynomial subtraction

There was a straightforward bug in the polynomial subtraction algorithm.

* add internal consistency check

The purpose of the internal consistency check is to prevent situations
like this from happening again.  Specifically, where we have no test
failures due to a bug in the translation from ASM to UASM (or similar).
Whilst unnecessary, this does provide an additional level of protection.