feat: add polynomial tests (#1235)
* extend polynomial tests

This significantly extends the range of testing for the polynomial
implementation.  At this stage, the tests are exposing some bugs.

* convert ArrayPoly to use AnySortedSet

This converts the implementation to use a sorted array, rather than just
an array.

* remove duplicates in NewAnySortedSet

This simply ensures duplicates are removed when a new set is being
created.

* fix agnostic.SplitPolynomial()

This was not correctly handling the case where the polynomial was zero,
leading to an unexpected panic.  This arose because polynomials are now
stricter about how they represent 0.