support "pub" modifier for functions
This adds support for a "pub" modifier on functions which is used to
specify whether or not the function is "externally visible".  This has a
somewhat odd meaning: non-externally visible functions are not available
in the generated Trace.java interface, nor are they initially visible in
the inspector.  At this time, that is their only meaning.