feat: support "pub" modifier for functions (#1236)
* support "pub" modifier for functions

This adds support for a "pub" modifier on functions which is used to
specify whether or not the function is "externally visible".  This has a
somewhat odd meaning: non-externally visible functions are not available
in the generated Trace.java interface, nor are they initially visible in
the inspector.  At this time, that is their only meaning.

* update some tests to use KOALABEAR

This just takes the opportunity to restore some tests which were
previously running on KOALABEAR.